pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "moviedemo"
version = "4.0"
dependencies = [
  "django==5.0.*",
  "dj-database-url",
  "goodconf[yaml]",
  "django-alive",
  "django-csp",
  "gunicorn",
  "psycopg",
  "python-json-logger",
  "sentry-sdk",
  "whitenoise",
  "wsgi-basic-auth",
  "djangorestframework",
]

[project.optional-dependencies]
dev = [
  "cogapp",
  "django-stubs",
  "mypy",
  "pip-tools",
  "readable-log-formatter",
  "ruff",
]

[tool.setuptools]
packages = { "find" = { } }
include-package-data = true

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "{{ project_name }}.settings"

[tool.ruff]
target-version = "py312"
exclude = ["migrations", ".venv"]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "ANN101",   # Missing Type Annotation for "self"
  "ANN401",   # Dynamically typed expressions (typing.Any) are disallowed in `**kwargs`"
  "ARG001",   # Unused function argument (request, ...)
  "ARG002",   # Unused method argument (*args, **kwargs)
  "D",        # Missing or badly formatted docstrings
  "E501",     # Let the formatter handle long lines
  "FBT",      # Flake Boolean Trap (don't use arg=True in functions)
  "RUF012",   # Mutable class attributes https://github.com/astral-sh/ruff/issues/5243

  "COM812",   # (ruff format) Checks for the absence of trailing commas
  "ISC001",   # (ruff format) Checks for implicitly concatenated strings on a single line
]

[tool.ruff.lint.extend-per-file-ignores]
# Also ignore `E402` in all `__init__.py` files.
"test_*.py" = [
  "S101",     # S101 Use of `assert` detected
  "PLR2004",  # Magic value used in comparison,
]
